backend/pyproject.toml
[tool.poetry]
name = "organize-mail-backend"
version = "0.1.0"
description = "Backend for organize-mail"
authors = ["you <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.100"
uvicorn = "^0.22"
google-api-python-client = "^2.0"
google-auth-oauthlib = "^1.0"
google-cloud-pubsub = "^3.0"
pydantic = "^2.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
